+++ src/main/java/net/tomasbot/mp/api/service/FileMetadataScanner.java
@@ -1,9 +1,10 @@
 package net.tomasbot.mp.api.service;
 
+import java.io.IOException;
 import org.jetbrains.annotations.NotNull;
 
 @FunctionalInterface
 public interface FileMetadataScanner<T> {
 
-  void scanFileMetadata(@NotNull T t) throws Throwable;
+  void scanFileMetadata(@NotNull T t) throws IOException;
 }