+++ src/main/java/net/tomasbot/mp/api/service/user/SetFavoriteException.java
@@ -1,0 +1,7 @@
+package net.tomasbot.mp.api.service.user;
+
+public class SetFavoriteException extends IllegalStateException {
+  public SetFavoriteException(Object o) {
+    super("Could not set favorite on: " + o);
+  }
+}