+++ src/main/java/net/tomasbot/mp/model/Editable.java
@@ -1,8 +1,11 @@
 package net.tomasbot.mp.model;
 
 import java.util.Set;
+import java.util.UUID;
 
 public interface Editable {
+
+  UUID getId();
 
   String getTitle();